Added cards, more animations, progress bar
Cards are going to need fixes as in the content of them. Maybe overall adjustments too but we'll see.

Animations are alright in my opinion.

Progress bar is maybe too small to instantly notice it, maybe making it bigger would help? I didn't want it to be overly large otherwise it would look a bit ugly