Updated header, added iphone, added photo library, added get in touch
I've updated the header to lose the register, login buttons once it contracts but they're overlapping. That minor issue can and will be fixed. Added a whatsapp showcase for the ai. added portfolio with some images from google. added get in touch. A lot for this one but a lot more to come as we discuss everything.